requirements.txt
flask
python-dotenv
google-cloud-aiplatform
google-auth-oauthlibpip install --upgrade google-api-python-client google-auth-httplib2 google-auth-oauthlib
google-api-python-client